Organize the project for hexagonal
