+++ src/domain/entities/boletoPayment.model.ts
