+++ src/domain/entities/customer.model.ts
