+++ src/domain/entities/pixPayment.model.ts
