+++ src/modules/customer.module.ts
@@ -1,6 +1,6 @@
 import { forwardRef, Module } from '@nestjs/common';
-import { CustomerController } from '../controllers/customer.controller';
-import { CustomerService } from '../services/customer.service';
+import { CustomerController } from '../application/controllers/customer.controller';
+import { CustomerService } from '../domain/services/customer.service';
 import { AccountModule } from './account.module';
 
 @Module({
